Return a null digest if the file doesn't exist
For cases where a symlink exists and is pointing to a file which does not exist, an error is thrown when you attempt to open that file because while the path you gave was valid, the file itself cannot be opened by the system.  This changes the behaviour so instead of throwing an error, an empty hex digest is returned instead.