Refactor how the directory is walked
This refactors how the directory is walked to make it slightly more clear how files are excluded from the hashing process. This was previously hidden within list comprehensions and guarded by the `ignore_hidden` flag (controlling one of two related ways). This breaks it out to make it more clear what is actually happening when `ignore_hidden` is passed in to the function. The actual behavioiur is not affected by this change, this is purely a refactor.